feat(OUT-1384): add improved preview modal with image title
